requirements.txt
python-telegram-bot==22.2
openpyxl==3.1.5
gspread==6.2.1
oauth2client==4.1.3
flask== 3.1.1